move rpc connection to module
